Add files via upload
Journal day4

Today I learned that this isn’t necessarily hard, but it also isn’t easy; and just because it isn’t easy, doesn’t mean it can’t be done. I also learned that it’s like learning a new spoken language - along with learning vocabulary, the learning person’s nervous system has to learn how to hear the language too. As the ear, cochlea, and acoustic nerve adjust, the brain gets better signals and can more accurately distill and discern. This is similar to coding in that part of learning includes learning how to understand and sort the information so the brain can accurately distill and learn.